Avoid "Marker id: 123456789 not found" errors
We have some code which runs asynchronously to add or clean up
error markers. In some cases this code runs after a project has
been closed or deleted, which results in errors. Guard these
marker access points with an IResource.isAccessible() check
(which performs both exists(), and for projects, isOpen()).

+++ eclipse/plugins/com.android.ide.eclipse.adt/src/com/android/ide/eclipse/adt/internal/project/BaseClasspathContainerInitializer.java
@@ -81,7 +81,7 @@ abstract class BaseClasspathContainerInitializer extends ClasspathContainerIniti
         } else {
             // no error, remove existing markers.
             try {
-                if (project.exists()) {
+                if (project.isAccessible()) {
                     project.deleteMarkers(markerType, true,
                             IResource.DEPTH_INFINITE);
                 }
@@ -92,8 +92,10 @@ abstract class BaseClasspathContainerInitializer extends ClasspathContainerIniti
                     @Override
                     protected IStatus run(IProgressMonitor monitor) {
                         try {
-                            project.deleteMarkers(markerType, true,
-                                    IResource.DEPTH_INFINITE);
+                            if (project.isAccessible()) {
+                                project.deleteMarkers(markerType, true,
+                                        IResource.DEPTH_INFINITE);
+                            }
                         } catch (CoreException e2) {
                             return e2.getStatus();
                         }
